Repair RVV opcode_fuse_t layout
Previous V-extension commits silently broke macro-op fusion. The first V
commit widened rv_insn_t.opcode from uint8_t to uint16_t (needed because
the V extension pushes fuse-opcode IDs past 255), but opcode_fuse_t.opcode
was left at uint8_t and the layout static_asserts were not updated.

After the widen, alignment padding moved rv_insn_t.opcode from byte 7
to byte 8. opcode_fuse_t still expected opcode at byte 7, and try_fuse_sequence's
memcpy(ir->fuse, ir, sizeof(opcode_fuse_t)) copied the padding byte into
fuse[i].opcode. Fused dispatch then ran on garbage, causing puzzle and
any fusion-eligible workload to enter undefined behavior (an infinite
loop in practice) under both interpreter and JIT. This was reproducible
on master + the V branch with EXT_V=n; the V state itself had nothing to
do with the symptom.

+++ src/t2c.c
@@ -238,7 +238,7 @@ static const void *dispatch_table[] = {
 #undef _
 };
 
-FORCE_INLINE bool t2c_insn_is_terminal(uint8_t opcode)
+FORCE_INLINE bool t2c_insn_is_terminal(uint16_t opcode)
 {
     switch (opcode) {
     case rv_insn_ecall:
